fix conflict and refactor
