+++ ui/.npmrc
@@ -1,0 +1,1 @@
+registry=https://skuberplus-dev.cloud.runsystem.site/registry
