--- ui/src/components/atoms/Drawer/index.ts
@@ -1,1 +1,0 @@
-export { Drawer } from "./Drawer";
