--- ui/src/components/atoms/Table/index.ts
@@ -1,2 +1,0 @@
-export { Table } from "./Table";
-export { TableEmpty, TableListEmpty } from "./TableEmpty";
