+++ ui/src/components/icons/AddIcon.tsx
@@ -1,5 +1,6 @@
 import { BaseIcon, BaseIconProps } from "./BaseIcon";
-import Icon from "../../assets/icons/add.svg";
+
+import Icon from "@/assets/icons/add.svg";
 
 export const AddIcon = (props: BaseIconProps) => {
   return <BaseIcon component={Icon} viewBox="0 0 16 16" {...props} />;
