+++ ui/src/components/icons/ArrowDownIcon.tsx
@@ -1,5 +1,6 @@
 import { BaseIcon, BaseIconProps } from "./BaseIcon";
-import Icon from "../../assets/icons/down.svg";
+
+import Icon from "@/assets/icons/down.svg";
 
 export const ArrowDownIcon = (props: BaseIconProps) => {
   return <BaseIcon size={16} component={Icon} viewBox="0 0 16 16" {...props} />;
