+++ ui/src/components/icons/CheckBoxIcon.tsx
@@ -1,5 +1,6 @@
 import { BaseIcon, BaseIconProps } from "./BaseIcon";
-import Icon from "../../assets/icons/checkboxicon.svg";
+
+import Icon from "@/assets/icons/checkboxicon.svg";
 
 export const CheckBoxIcon = (props: BaseIconProps) => {
   return <BaseIcon component={Icon} size={12} viewBox="0 0 13 12" {...props} />;
