+++ ui/src/components/icons/InfoIcon.tsx
@@ -1,5 +1,6 @@
 import { BaseIcon, BaseIconProps } from "./BaseIcon";
-import Icon from "../../assets/icons/info.svg";
+
+import Icon from "@/assets/icons/info.svg";
 
 export const InfoIcon = (props: BaseIconProps) => {
   return <BaseIcon component={Icon} viewBox="0 0 16 16" {...props} />;
