+++ ui/src/components/pages/WorkloadDetail.tsx
@@ -2,12 +2,13 @@ import { useState } from "react";
 
 import { Button } from "@skuber/components";
 
-import { Drawer } from "../atoms/Drawer";
-import { ClosePort } from "./components/pages/detail/ClosePort";
-import { OpenPort } from "./components/pages/detail/OpenPort";
-import { PolicyApplication } from "./components/pages/detail/PolicyApplication";
-import { WorkloadSummary } from "./components/pages/detail/WorkloadSummary";
-import { WorkloadTabs } from "./components/pages/detail/WorkloadTabs";
+import { ClosePort } from "./detail/ClosePort";
+import { OpenPort } from "./detail/OpenPort";
+import { PolicyApplication } from "./detail/PolicyApplication";
+import { WorkloadSummary } from "./detail/WorkloadSummary";
+import { WorkloadTabs } from "./detail/WorkloadTabs";
+
+import { Drawer } from "@/components/atoms/Drawer";
 
 export const WorkloadDetail = () => {
   const [isWorkloadDetailOpened, setIsWorkloadDetailOpened] = useState(false);
