+++ ui/src/components/pages/detail/PortDetail.tsx
