+++ ui/src/components/pages/detail/data.ts
