+++ ui/vite.config.ts
@@ -1,3 +1,5 @@
+import path from "path";
+
 import react from "@vitejs/plugin-react-swc";
 import { defineConfig } from "vite";
 import svgr from "vite-plugin-svgr";
@@ -10,4 +12,9 @@ export default defineConfig({
       include: "**/*.svg",
     }),
   ],
+  resolve: {
+    alias: {
+      "@": path.resolve(__dirname, "./src"),
+    },
+  },
 });
